+++ src/main/java/com/apiround/greenhub/repository/RecipeIngredientRepository.java
@@ -4,6 +4,6 @@ import com.apiround.greenhub.entity.RecipeIngredient;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface RecipeIngredientRepository extends JpaRepository<RecipeIngredient, Long> {
-    List<RecipeIngredient> findByRecipeRecipeIdOrderByLineNoAsc(Long recipeId);
+public interface RecipeIngredientRepository extends JpaRepository<RecipeIngredient, Integer> {
+    List<RecipeIngredient> findByRecipeRecipeIdOrderByLineNoAsc(Integer recipeId);
 }
